# core/generated/swagger-ui/schema-registry-api-spec.yaml
...
swagger: "2.0"
info:
  version: "v1"
  title: "Confluent Schema Registry"
schemes:
- "http"
- "https"
paths:
...
  /schemas/ids/{id}:
    get:
      summary: "Get the schema string identified by the input ID."
      description: ""
      operationId: "getSchema"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "id"
        in: "path"
        description: "Globally unique identifier of the schema"
        required: true
        type: "integer"
        format: "int32"
      - name: "fetchMaxId"
        in: "query"
        required: false
        type: "boolean"
        default: false
      responses:
        200:
          description: "successful operation"
          schema:
            $ref: "#/definitions/SchemaString"
        404:
          description: "Error code 40403 -- Schema not found\n"
        500:
          description: "Error code 50001 -- Error in the backend data store\n"
...
  /subjects:
    get:
      summary: "Get a list of registered subjects."
      description: ""
      operationId: "list"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters: []
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              type: "string"
            uniqueItems: true
        500:
          description: "Error code 50001 -- Error in the backend datastore"
  /subjects/{subject}:
    post:
      summary: "Check if a schema has already been registered under the specified\
        \ subject. If so, this returns the schema string along with its globally unique\
        \ identifier, its version under this subject and the subject name."
      description: ""
      operationId: "lookUpSchemaUnderSubject"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "subject"
        in: "path"
        description: "Subject under which the schema will be registered"
        required: true
        type: "string"
      - name: "deleted"
        in: "query"
        required: false
        type: "boolean"
      - in: "body"
        name: "body"
        description: "Schema"
        required: true
        schema:
          $ref: "#/definitions/RegisterSchemaRequest"
      responses:
        404:
          description: "Error code 40401 -- Subject not found\nError code 40403 --\
            \ Schema not found"
        500:
          description: "Internal server error"
          schema:
            $ref: "#/definitions/Schema"
    delete:
      summary: "Deletes the specified subject and its associated compatibility level\
        \ if registered. It is recommended to use this API only when a topic needs\
        \ to be recycled or in development environment."
      description: ""
      operationId: "deleteSubject"
      consumes:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json"
      - "application/json"
      - "application/octet-stream"
      produces:
      - "application/vnd.schemaregistry.v1+json"
      - "application/vnd.schemaregistry+json; qs=0.9"
      - "application/json; qs=0.5"
      parameters:
      - name: "subject"
        in: "path"
        description: "the name of the subject"
        required: true
        type: "string"
      responses:
        200:
          description: "successful operation"
          schema:
            type: "array"
            items:
              type: "integer"
              format: "int32"
        404:
          description: "Error code 40401 -- Subject not found"
        500:
          description: "Error code 50001 -- Error in the backend datastore"
...
definitions:
...
  RegisterSchemaRequest:
    type: "object"
    properties:
      version:
        type: "integer"
        format: "int32"
      id:
        type: "integer"
        format: "int32"
      schemaType:
        type: "string"
      schema:
        type: "string"
...
  Schema:
    type: "object"
    properties:
      subject:
        type: "string"
      version:
        type: "integer"
        format: "int32"
      id:
        type: "integer"
        format: "int32"
      schema:
        type: "string"
      schemaType:
        type: "string"
  SchemaString:
    type: "object"
    properties:
      schema:
        type: "string"
        description: "Schema string identified by the ID"
      maxId:
        type: "integer"
        format: "int32"
...